update usage doc to reflect all fitted sizes
